Add integration tests for story submission.
This added an integration helper for sigining in.

This also breaks the navbar by moving the "Story Submission" out at the
end so that it gets seen by phantomjs. This needs to get fixed later.

This also adds a bunch of `id` tags to the navbar and to the story view
page to ease testing.